pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
wand = "^0.6.6"
biopython = "^1.78"
pil-compat = "^1.0.0"
pytesseract = "^0.3.7"
pdf2image = "^1.15.1"
PyPDF2 = "^1.26.0"
[tool.poetry.dev-dependencies]
